pyproject.toml
[tool.poetry]
name = "josim_tools"
version = "0.1.0"
description = ""
authors = ["Paul le Roux <pleroux0@outlook.com>"]
license = "BSD-2-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
numpy = "^1.16"
matplotlib = "^3.0.3"
jsonschema = "^3.0.1"
toml = "^0.10.0"
attrs = "^19.1.0"
scipy = "^1.2.1"
typing_extensions = "^3.7.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
'josim-tools' = 'josim_tools.tools:run'

[tool.black]
line-length = 88
target_version = ['py36']

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
